build(warnings): zero host compile-warning baseline (Phase 2, step 1)
Box gcc 10.3.1 host build went from 40 compiler warning lines to 0; the
only remaining warning output is the DDR manifest alias report (own fix).

Production (explicit casts documenting the existing implicit conversions;
bit-identical on ARM and x86 host, no behavior change):
- multiregion_context.c: 10 flash-address casts (void*)TIER[12]_ADDRS[x]
  now go through uintptr_t (host is 64-bit; ARM was already clean); 6
  sign-conversion sites cast explicitly (uint8_t loop index -> int8_t
  slot x2, int8_t -> uint8_t active_slot x2, uint8_t -> int8_t finder
  return, int8_t ChannelsDatarate -> uint8_t datarate).
- atgm336h.c: checksum XOR widens the (signed on x86) char explicitly;
  memcpy length widened after its existing clamp.

Harness:
- 7 slurp() copies: CRLF strip loop compares size_t r < (size_t)n.
- test_multiregion.c: 4 fake-flash address casts through uintptr_t.
- test_review_20260812.c: drop dead count_occurrences/in_function helpers.
- test_main.c: one gcc-9.2-only sign-conversion at the wire-CRC check.

Verified: box clean all = all suites green (3 BASELINE-OK characterization
gates), 0 compiler warnings; local gcc 9.2 builds all 24 targets
warning-free; edited suites green (multiregion 98, stability 49,
review0812 6, dr 51, sp0813 73, txadapter 11, gnssfix 10, gpsloss 9).

+++ Core/Src/atgm336h.c
@@ -984,7 +984,7 @@ uint8_t GNSS_CalculateChecksum(const char *sentence)
   /* Calculate XOR of all characters until * or end */
   while (*p != '\0' && *p != '*')
   {
-    checksum ^= *p;
+    checksum ^= (uint8_t)*p;
     p++;
   }
 
@@ -1142,7 +1142,7 @@ static bool GNSS_GetToken(const char *sentence, int index, char *buffer, int max
   if (len < 0) len = 0;
   if (len >= max_len) len = max_len - 1;
 
-  memcpy(buffer, start, len);
+  memcpy(buffer, start, (size_t)len);
   buffer[len] = '\0';
 
   return true;
